service/Services/entities/src/main/java/org/infy/entities/triggerinputs/SubModule.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.entities.triggerinputs;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store sub module details
 * 
 * @author Infosys
 *
 */
public class SubModule {

	@SerializedName("moduleName")
	@Expose
	private String moduleName;
	@SerializedName("defaultModule")
	@Expose
	private String defaultModule;

	public String getModuleName() {
		return moduleName;
	}

	public void setModuleName(String moduleName) {
		this.moduleName = moduleName;
	}

	public String getDefaultModule() {
		return defaultModule;
	}

	public void setDefaultModule(String defaultModule) {
		this.defaultModule = defaultModule;
	}

}
